Move (rename) 'api/generated' to 'shared/api'
