+++ src/shared/api/core/OpenAPI.ts
