+++ src/shared/api/models/DocAttribute.ts
