+++ src/shared/api/models/DocAttributeDto.ts
