+++ src/shared/api/models/DocAttributeValues.ts
