+++ src/shared/api/models/DocStatisticDTO.ts
