+++ src/shared/api/models/UserRatingDto.ts
